+++ core-ng/src/main/java/core/framework/impl/web/management/PropertyController.java
@@ -5,7 +5,8 @@ import core.framework.web.Controller;
 import core.framework.web.Request;
 import core.framework.web.Response;
 
-import java.util.List;
+import java.util.Set;
+import java.util.TreeSet;
 
 /**
  * @author neo
@@ -25,10 +26,10 @@ public class PropertyController implements Controller {
 
     String text() {
         StringBuilder builder = new StringBuilder();
-        List<PropertyManager.PropertyEntry> entries = propertyManager.entries();
-        for (PropertyManager.PropertyEntry entry : entries) {
-            if (entry.source != PropertyManager.PropertySource.PROPERTY_FILE) builder.append("# ").append(entry.key).append(" is overridden by ").append(entry.source).append('\n');
-            builder.append(entry.key).append('=').append(entry.maskedValue()).append('\n');
+        Set<String> keys = new TreeSet<>(propertyManager.properties.keys());   // sort by key
+        for (String key : keys) {
+            String value = propertyManager.property(key).orElse("");
+            builder.append(key).append('=').append(propertyManager.maskValue(key, value)).append('\n');
         }
         return builder.toString();
     }