+++ core-ng/src/test/java/core/framework/impl/module/PropertyManagerTest.java
@@ -30,14 +30,14 @@ class PropertyManagerTest {
     }
 
     @Test
-    void maskedValue() {
-        PropertyManager.PropertyEntry entry = new PropertyManager.PropertyEntry("sys.jdbc.password", "password", PropertyManager.PropertySource.PROPERTY_FILE);
-        assertNotEquals("password", entry.maskedValue());
+    void maskValue() {
+        assertNotEquals("password", propertyManager.maskValue("sys.jdbc.password", "password"));
+        assertNotEquals("secret", propertyManager.maskValue("app.key.secret", "secret"));
+        assertEquals("user", propertyManager.maskValue("sys.jdbc.user", "user"));
+    }
 
-        entry = new PropertyManager.PropertyEntry("app.key.secret", "secret", PropertyManager.PropertySource.PROPERTY_FILE);
-        assertNotEquals("secret", entry.maskedValue());
-
-        entry = new PropertyManager.PropertyEntry("sys.jdbc.user", "user", PropertyManager.PropertySource.PROPERTY_FILE);
-        assertEquals("user", entry.maskedValue());
+    @Test
+    void envVarName() {
+        assertEquals("SYS_KAFKA_URI", propertyManager.envVarName("sys.kafka.uri"));
     }
 }