+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Change log
+### 5.2.8 (12/20/2017)
+* property: when override property with env var, convert key with upper case and replace '.' with '_', e.g. "sys.kafka.uri" to "SYS_KAFKA_URI", 
+    due to dot is not supported by POSIX, especially not supported by alpine 3.6, https://bugs.alpinelinux.org/issues/7344 
+
 ### 5.2.7 (12/11/2017 - 12/19/2017)
 * property: allow all the properties can be override by env var (mount by kubernetes/docker) or system property (via -Dkey=value)
 * search: update es to 6.1.0, disable zen for integration test